# Stack optimization for eytzinger_walk()