`eytzinger_walk()` is used to determine data insertion order. It can be made on average 5% faster if we collapse arguments. Because this function is called recursively, the number of the arguments dictates the stack size and the number of `push`/`pop` instruction in prolog/epilog required to maintain CPU state.

<details>
  <summary>Results against main branch</summary>
  
```console
$  cargo bench --bench search_comparison -- "Construction u(.*)/ordsearch/" --baseline master
Construction u8/ordsearch/8
                        time:   [84.249 ns 84.928 ns 85.623 ns]
                        change: [-8.7782% -5.9025% -3.5420%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/16
                        time:   [123.35 ns 124.15 ns 125.05 ns]
                        change: [-9.7299% -6.9626% -4.0815%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/32
                        time:   [243.44 ns 245.05 ns 246.74 ns]
                        change: [-10.261% -8.5700% -6.9796%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/64
                        time:   [520.70 ns 525.53 ns 531.17 ns]
                        change: [-4.7382% -3.4615% -1.9882%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/128
                        time:   [1.1775 µs 1.1874 µs 1.1996 µs]
                        change: [+0.9845% +2.1546% +3.2931%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/256
                        time:   [1.6764 µs 1.6901 µs 1.7057 µs]
                        change: [-6.8188% -4.9195% -3.0613%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/512
                        time:   [2.6154 µs 2.6294 µs 2.6440 µs]
                        change: [-4.3197% -3.0121% -1.7527%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/1024
                        time:   [4.3218 µs 4.3506 µs 4.3815 µs]
                        change: [-5.5349% -4.5061% -3.4314%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/2048
                        time:   [7.8819 µs 7.9399 µs 7.9963 µs]
                        change: [-6.2673% -4.8777% -3.5181%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/4069
                        time:   [15.387 µs 15.706 µs 16.061 µs]
                        change: [-4.6401% -3.0945% -1.5000%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/8192
                        time:   [29.289 µs 29.531 µs 29.784 µs]
                        change: [-7.6373% -5.7003% -3.9959%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/16384
                        time:   [56.927 µs 57.276 µs 57.664 µs]
                        change: [-8.3792% -6.5118% -4.8294%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/32768
                        time:   [113.87 µs 114.63 µs 115.57 µs]
                        change: [-6.7090% -4.6875% -2.3905%] (p = 0.00 < 0.05)
--
Construction u8/ordsearch/65536
                        time:   [239.81 µs 242.05 µs 244.41 µs]
                        change: [-11.427% -8.6849% -6.1105%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/8
                        time:   [85.498 ns 86.250 ns 87.009 ns]
                        change: [-7.2359% -5.0428% -2.8818%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/16
                        time:   [138.98 ns 139.89 ns 140.89 ns]
                        change: [-5.9272% -3.2772% -0.5964%] (p = 0.02 < 0.05)
--
Construction u16/ordsearch/32
                        time:   [254.38 ns 256.08 ns 257.90 ns]
                        change: [-6.9991% -5.4902% -3.9544%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/64
                        time:   [556.34 ns 560.09 ns 564.05 ns]
                        change: [-7.3324% -5.4013% -3.6549%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/128
                        time:   [1.2144 µs 1.2379 µs 1.2781 µs]
                        change: [-0.2762% +1.0381% +2.5999%] (p = 0.19 > 0.05)
--
Construction u16/ordsearch/256
                        time:   [2.8204 µs 2.8349 µs 2.8490 µs]
                        change: [-0.2461% +0.6366% +1.5465%] (p = 0.17 > 0.05)
--
Construction u16/ordsearch/512
                        time:   [5.7926 µs 5.8352 µs 5.8802 µs]
                        change: [+0.3582% +1.4187% +2.4477%] (p = 0.01 < 0.05)
--
Construction u16/ordsearch/1024
                        time:   [11.921 µs 11.995 µs 12.071 µs]
                        change: [+0.3893% +1.4193% +2.4450%] (p = 0.01 < 0.05)
--
Construction u16/ordsearch/2048
                        time:   [25.684 µs 26.233 µs 27.041 µs]
                        change: [+2.8825% +5.4044% +9.3368%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/4069
                        time:   [65.103 µs 65.728 µs 66.358 µs]
                        change: [-6.4551% -3.7589% -0.9922%] (p = 0.01 < 0.05)
--
Construction u16/ordsearch/8192
                        time:   [168.62 µs 169.91 µs 171.38 µs]
                        change: [-4.7784% -3.4312% -1.9147%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/16384
                        time:   [386.17 µs 388.71 µs 391.55 µs]
                        change: [-5.9679% -4.8689% -3.8590%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/32768
                        time:   [865.06 µs 873.90 µs 883.81 µs]
                        change: [-6.5229% -5.2712% -3.8514%] (p = 0.00 < 0.05)
--
Construction u16/ordsearch/65536
                        time:   [1.0266 ms 1.0348 ms 1.0435 ms]
                        change: [-7.5373% -6.0600% -4.7010%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/8
                        time:   [88.761 ns 89.330 ns 89.939 ns]
                        change: [-8.7846% -5.8078% -2.9699%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/16
                        time:   [142.06 ns 143.21 ns 144.42 ns]
                        change: [-9.9996% -7.1037% -4.6230%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/32
                        time:   [256.30 ns 259.19 ns 262.53 ns]
                        change: [-8.8452% -7.1138% -5.3164%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/64
                        time:   [522.50 ns 525.74 ns 529.06 ns]
                        change: [-11.809% -10.486% -9.1334%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/128
                        time:   [1.2079 µs 1.2185 µs 1.2310 µs]
                        change: [-10.150% -8.5346% -6.8275%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/256
                        time:   [2.5803 µs 2.5977 µs 2.6168 µs]
                        change: [-4.0172% -2.6726% -1.3984%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/512
                        time:   [5.4646 µs 5.5179 µs 5.5804 µs]
                        change: [-2.8120% -1.6680% -0.4634%] (p = 0.01 < 0.05)
--
Construction u32/ordsearch/1024
                        time:   [11.187 µs 11.293 µs 11.415 µs]
                        change: [-2.8312% -1.4828% -0.1428%] (p = 0.03 < 0.05)
--
Construction u32/ordsearch/2048
                        time:   [24.739 µs 24.960 µs 25.182 µs]
                        change: [-4.9350% -3.3332% -1.7580%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/4069
                        time:   [63.631 µs 64.236 µs 64.890 µs]
                        change: [-17.863% -16.670% -15.516%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/8192
                        time:   [172.71 µs 173.99 µs 175.33 µs]
                        change: [-8.9083% -7.9447% -6.9264%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/16384
                        time:   [432.75 µs 437.49 µs 442.42 µs]
                        change: [-8.4724% -6.9994% -5.4725%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/32768
                        time:   [879.06 µs 885.82 µs 892.82 µs]
                        change: [-7.6233% -6.1567% -4.7446%] (p = 0.00 < 0.05)
--
Construction u32/ordsearch/65536
                        time:   [1.7848 ms 1.7956 ms 1.8068 ms]
                        change: [-5.5351% -4.4413% -3.4155%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/8
                        time:   [99.611 ns 100.74 ns 102.02 ns]
                        change: [-8.9870% -5.4388% -1.0474%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/16
                        time:   [151.83 ns 153.63 ns 155.86 ns]
                        change: [-14.974% -12.221% -9.7487%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/32
                        time:   [273.98 ns 276.02 ns 278.11 ns]
                        change: [-17.333% -14.587% -12.384%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/64
                        time:   [608.59 ns 613.40 ns 618.60 ns]
                        change: [-8.3341% -7.0051% -5.7471%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/128
                        time:   [1.1401 µs 1.1540 µs 1.1700 µs]
                        change: [-7.7534% -6.4648% -5.0601%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/256
                        time:   [2.5511 µs 2.5695 µs 2.5877 µs]
                        change: [-9.1214% -7.8099% -6.3207%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/512
                        time:   [5.2646 µs 5.3861 µs 5.5748 µs]
                        change: [-9.9701% -8.3999% -6.5880%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/1024
                        time:   [11.058 µs 11.157 µs 11.266 µs]
                        change: [-9.4877% -7.9111% -6.3850%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/2048
                        time:   [24.637 µs 24.917 µs 25.220 µs]
                        change: [-3.5525% -1.8763% -0.1549%] (p = 0.03 < 0.05)
--
Construction u64/ordsearch/4069
                        time:   [73.283 µs 73.864 µs 74.509 µs]
                        change: [+7.4984% +9.5032% +11.394%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/8192
                        time:   [194.68 µs 197.26 µs 200.01 µs]
                        change: [-6.4736% -2.6044% +0.9249%] (p = 0.20 > 0.05)
--
Construction u64/ordsearch/16384
                        time:   [444.81 µs 450.04 µs 455.54 µs]
                        change: [-4.7211% -2.9710% -1.1609%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/32768
                        time:   [898.10 µs 905.75 µs 913.63 µs]
                        change: [-3.6622% -2.3154% -1.0787%] (p = 0.00 < 0.05)
--
Construction u64/ordsearch/65536
                        time:   [1.8666 ms 1.8803 ms 1.8944 ms]
                        change: [-0.7021% +0.9236% +2.4003%] (p = 0.24 > 0.05)
--
Construction u128/ordsearch/8
                        time:   [122.21 ns 123.49 ns 124.79 ns]
                        change: [-9.3002% -5.3464% -1.5456%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/16
                        time:   [207.03 ns 209.20 ns 211.50 ns]
                        change: [-9.5308% -6.7113% -3.9545%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/32
                        time:   [441.82 ns 451.90 ns 461.13 ns]
                        change: [-7.6109% -5.6748% -3.6842%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/64
                        time:   [747.81 ns 754.82 ns 762.28 ns]
                        change: [-6.9463% -5.2452% -3.6448%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/128
                        time:   [1.5557 µs 1.5699 µs 1.5842 µs]
                        change: [-2.4730% -0.5468% +2.2033%] (p = 0.69 > 0.05)
--
Construction u128/ordsearch/256
                        time:   [3.5852 µs 3.6171 µs 3.6489 µs]
                        change: [-1.1552% +0.2444% +1.7460%] (p = 0.75 > 0.05)
--
Construction u128/ordsearch/512
                        time:   [7.5056 µs 7.6160 µs 7.7370 µs]
                        change: [-2.8339% -0.4727% +1.8323%] (p = 0.69 > 0.05)
--
Construction u128/ordsearch/1024
                        time:   [15.812 µs 15.991 µs 16.169 µs]
                        change: [-7.7661% -5.4151% -3.2638%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/2048
                        time:   [34.237 µs 34.564 µs 34.885 µs]
                        change: [-12.619% -10.813% -9.1679%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/4069
                        time:   [99.762 µs 101.56 µs 103.55 µs]
                        change: [-24.450% -19.485% -14.075%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/8192
                        time:   [245.10 µs 248.88 µs 252.97 µs]
                        change: [-26.617% -22.736% -18.894%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/16384
                        time:   [567.50 µs 575.50 µs 583.70 µs]
                        change: [-15.033% -12.814% -10.617%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/32768
                        time:   [1.1453 ms 1.1563 ms 1.1675 ms]
                        change: [-10.733% -9.1753% -7.6157%] (p = 0.00 < 0.05)
--
Construction u128/ordsearch/65536
                        time:   [2.2746 ms 2.2971 ms 2.3211 ms]
                        change: [-7.3149% -5.7216% -4.0768%] (p = 0.00 < 0.05)

```
</details>